+++ tests/wps_restapi/test_jobs.py
@@ -1609,9 +1609,7 @@ class WpsRestApiJobsTest(JobUtils):
             assert resp.content_type == ContentType.APP_JSON
             assert resp.json["cause"]["schema"] == {"const": sd.Execute._schema}
             assert resp.json["cause"]["name"] == "Content-Schema"
-            assert resp.json["type"] == (
-                "http://www.opengis.net/def/exceptions/ogcapi-processes-4/1.0/unsupported-schema"
-            )
+            assert resp.json["type"] == sd.OGC_API_PROC_PART4_EXC_UNSUPPORTED_SCHEMA_URI
 
     def test_job_results_errors(self):
         """
@@ -2556,7 +2554,7 @@ class WpsRestApiJobsTest(JobUtils):
         body = {"inputs": {"test": 400}}
         resp = self.app.patch_json(path, params=body, headers=self.json_headers, expect_errors=True)
         assert resp.status_code == 423
-        assert resp.json["type"] == "http://www.opengis.net/def/exceptions/ogcapi-processes-4/1.0/locked"
+        assert resp.json["type"] == sd.OGC_API_PROC_PART4_EXC_LOCKED_URI
 
     @pytest.mark.oap_part4
     def test_job_update_unsupported_media_type(self):
@@ -2568,9 +2566,7 @@ class WpsRestApiJobsTest(JobUtils):
         resp = self.app.patch(path, params="data", expect_errors=True)
         assert resp.status_code == 415
         assert resp.content_type == ContentType.APP_JSON
-        assert resp.json["type"] == (
-            "http://www.opengis.net/def/exceptions/ogcapi-processes-4/1.0/unsupported-media-type"
-        )
+        assert resp.json["type"] == sd.OGC_API_PROC_PART4_EXC_UNSUPPORTED_MEDIA_TYPE_URI
 
     @pytest.mark.oap_part4
     def test_job_update_response_contents(self):
@@ -3856,3 +3852,32 @@ def test_get_job_status_profile_invalid(accept_type, accept_profile):
     request = MockedRequest(headers={"Accept": accept_type, "Accept-Profile": accept_profile})
     with pytest.raises(HTTPBadRequest):
         get_job_status_schema(request)
+
+
+@pytest.mark.parametrize(
+    ["accept_profile", "query_profile", "expected_profile"],
+    [
+        (None, JobStatusProfileSchema.OGC.upper(), JobStatusProfileSchema.OGC),
+        (None, JobStatusProfileSchema.OGC.lower(), JobStatusProfileSchema.OGC),
+        (None, JobStatusProfileSchema.WPS.upper(), JobStatusProfileSchema.WPS),
+        (None, JobStatusProfileSchema.WPS.lower(), JobStatusProfileSchema.WPS),
+        (JobStatusProfileSchema.OGC.upper(), None, JobStatusProfileSchema.OGC),
+        (JobStatusProfileSchema.OGC.lower(), None, JobStatusProfileSchema.OGC),
+        (JobStatusProfileSchema.WPS.upper(), None, JobStatusProfileSchema.WPS),
+        (JobStatusProfileSchema.WPS.lower(), None, JobStatusProfileSchema.WPS),
+        (None, sd.OGC_API_PROC_PROFILE_JOB_DESC_URI.replace("http://", "https://"), JobStatusProfileSchema.OGC),
+        (sd.OGC_API_PROC_PROFILE_JOB_DESC_URI.replace("http://", "https://"), None, JobStatusProfileSchema.OGC),
+        (None, sd.OGC_API_PROC_PROFILE_JOB_DESC_URI.replace("https://", "http://"), JobStatusProfileSchema.OGC),
+        (sd.OGC_API_PROC_PROFILE_JOB_DESC_URI.replace("https://", "http://"), None, JobStatusProfileSchema.OGC),
+    ]
+)
+def test_get_job_status_profile_valid(accept_profile, query_profile, expected_profile):
+    """
+    Test valid combinations of :term:`Job` status :term:`Profile`.
+    """
+    queries = {"profile": query_profile} if query_profile else {}
+    headers = {"Accept": ContentType.APP_JSON}
+    headers.update({"Accept-Profile": accept_profile} if accept_profile else {})
+    request = MockedRequest(params=queries, headers=headers)
+    profile, _ = get_job_status_schema(request)
+    assert profile == expected_profile