+++ weaver/wps_restapi/swagger_definitions.py
@@ -12,12 +12,12 @@ definitions generated on the exposed endpoints (JSON and Swagger UI).
 The definitions are also employed to generate the `OpenAPI` definitions reported in the documentation published
 on `Weaver`'s `ReadTheDocs` page.
 """
-
 # pylint: disable=C0103,invalid-name
 # pylint: disable=E0241,duplicate-bases
 
 import datetime
 import inspect
+import json
 import os
 import re
 from copy import copy
@@ -128,6 +128,7 @@ from weaver.wps_restapi.colander_extras import (
     AnyOfKeywordSchema,
     BoundedRange,
     CommaSeparated,
+    DelimitedStringOneOf,
     EmptyMappingSchema,
     ExpandStringList,
     ExtendedBoolean as Boolean,
@@ -148,7 +149,6 @@ from weaver.wps_restapi.colander_extras import (
     SchemeURL,
     SemanticVersion,
     StrictMappingSchema,
-    StringOneOf,
     StringRange,
     XMLObject
 )
@@ -156,7 +156,7 @@ from weaver.wps_restapi.constants import ConformanceCategory
 from weaver.wps_restapi.patches import WeaverService as Service  # warning: don't use 'cornice.Service'
 
 if TYPE_CHECKING:
-    from typing import Any, Dict, List, Type, Union
+    from typing import Any, Dict, Iterable, List, Tuple, Type, Union
     from typing_extensions import TypedDict
 
     from pygeofilter.ast import AstType as FilterAstType
@@ -183,7 +183,8 @@ CWL_REPO_URL = "https://github.com/common-workflow-language"
 CWL_SCHEMA_BRANCH = "v1.2.1"
 CWL_SCHEMA_PATH = "json-schema/cwl.yaml"
 CWL_SCHEMA_REPO = f"https://raw.githubusercontent.com/common-workflow-language/cwl-{CWL_VERSION}"
-CWL_SCHEMA_URL = f"{CWL_SCHEMA_REPO}/{CWL_SCHEMA_BRANCH}/{CWL_SCHEMA_PATH}"
+CWL_SCHEMA_SOURCE_URL = f"{CWL_SCHEMA_REPO}/{CWL_SCHEMA_BRANCH}/{CWL_SCHEMA_PATH}"
+CWL_SCHEMA_URL = f"https://w3id.org/cwl/{CWL_VERSION}/cwl-json-schema.yaml"
 CWL_BASE_URL = "https://www.commonwl.org"
 CWL_SPEC_URL = f"{CWL_BASE_URL}/#Specification"
 CWL_USER_GUIDE_URL = f"{CWL_BASE_URL}/user_guide"
@@ -235,21 +236,43 @@ OGC_API_PROC_BBOX_SCHEMA = f"{OGC_API_PROC_PART1_SCHEMAS}/bbox.yaml"
 OGC_API_PROC_BBOX_FORMAT = "ogc-bbox"  # equal CRS:84 and EPSG:4326, equivalent to WGS84 with swapped lat-lon order
 OGC_API_PROC_BBOX_CRS = ogc_def.OGC_DEF_CRS_CRS84_URI
 
-OGC_API_PROC_REL_EXCEPTIONS_URI = "http://www.opengis.net/def/rel/ogc/1.0/exceptions"
-OGC_API_PROC_REL_EXECUTE_URI = "http://www.opengis.net/def/rel/ogc/1.0/execute"
-OGC_API_PROC_REL_PROCESSES_URI = "http://www.opengis.net/def/rel/ogc/1.0/processes"
-OGC_API_PROC_REL_PROCESS_DESC_URI = "http://www.opengis.net/def/rel/ogc/1.0/process-desc"
-OGC_API_PROC_REL_JOB_RESULTS_URI = "http://www.opengis.net/def/rel/ogc/1.0/results"
-OGC_API_PROC_REL_JOB_LIST_URI = "http://www.opengis.net/def/rel/ogc/1.0/job-list"
-OGC_API_PROC_REL_JOB_LOG_URI = "http://www.opengis.net/def/rel/ogc/1.0/log"
+OGC_API_PROC_PART1_EXC_BASE_URI = "https://www.opengis.net/def/exceptions/ogcapi-processes-1/1.0"
+OGC_API_PROC_PART1_EXC_INVALID_PARAMETER_URI = f"{OGC_API_PROC_PART1_EXC_BASE_URI}/invalid-parameter"
+OGC_API_PROC_PART1_EXC_NO_SUCH_PROCESS_URI = f"{OGC_API_PROC_PART1_EXC_BASE_URI}/no-such-process"
+OGC_API_PROC_PART1_EXC_NO_SUCH_JOB_URI = f"{OGC_API_PROC_PART1_EXC_BASE_URI}/no-such-job"
+OGC_API_PROC_PART1_EXC_NO_SUCH_OUTPUT_URI = f"{OGC_API_PROC_PART1_EXC_BASE_URI}/no-such-output"
+OGC_API_PROC_PART1_EXC_RESULT_NOT_AVAILABLE_URI = f"{OGC_API_PROC_PART1_EXC_BASE_URI}/result-not-available"
+OGC_API_PROC_PART1_EXC_RESULT_NOT_READY_URI = f"{OGC_API_PROC_PART1_EXC_BASE_URI}/result-not-ready"
 
-OGC_API_PROC_PROFILE_PROC_DESC_URI = "http://www.opengis.net/def/profile/OGC/0/ogc-process-description"
-OGC_API_PROC_PROFILE_PROC_LIST_URI = "http://www.opengis.net/def/profile/OGC/0/ogc-process-list"
-OGC_API_PROC_PROFILE_EXECUTE_URI = "http://www.opengis.net/def/profile/OGC/0/ogc-execute-request"
-OGC_API_PROC_PROFILE_RESULTS_URI = "http://www.opengis.net/def/profile/OGC/0/ogc-results"
-OGC_API_PROC_PROFILE_JOB_DESC_URI = "http://www.opengis.net/def/profile/OGC/0/job-description"
-OGC_API_PROC_PROFILE_JOB_LIST_URI = "http://www.opengis.net/def/profile/OGC/0/jobs-list"
+OGC_API_PROC_PART2_EXC_BASE_URI = "https://www.opengis.net/def/exceptions/ogcapi-processes-2/1.0"
+OGC_API_PROC_PART2_EXC_IMMUTABLE_PROCESS_URI = f"{OGC_API_PROC_PART2_EXC_BASE_URI}/immutable-process"
+OGC_API_PROC_PART2_EXC_DUPLICATED_PROCESS_URI = f"{OGC_API_PROC_PART2_EXC_BASE_URI}/duplicated-process"
+OGC_API_PROC_PART2_EXC_UNSUPPORTED_MEDIA_TYPE_URI = f"{OGC_API_PROC_PART2_EXC_BASE_URI}/unsupported-media-type"
 
+OGC_API_PROC_PART4_EXC_BASE_URI = "https://www.opengis.net/def/exceptions/ogcapi-processes-4/1.0"
+OGC_API_PROC_PART4_EXC_LOCKED_URI = f"{OGC_API_PROC_PART4_EXC_BASE_URI}/locked"
+OGC_API_PROC_PART4_EXC_UNSUPPORTED_MEDIA_TYPE_URI = f"{OGC_API_PROC_PART4_EXC_BASE_URI}/unsupported-media-type"
+OGC_API_PROC_PART4_EXC_UNSUPPORTED_SCHEMA_URI = f"{OGC_API_PROC_PART4_EXC_BASE_URI}/unsupported-schema"
+
+OGC_API_PROC_REL_BASE_URI = "https://www.opengis.net/def/rel/ogc/1.0"
+OGC_API_PROC_REL_EXCEPTIONS_URI = f"{OGC_API_PROC_REL_BASE_URI}/exceptions"
+OGC_API_PROC_REL_EXECUTE_URI = f"{OGC_API_PROC_REL_BASE_URI}/execute"
+OGC_API_PROC_REL_PROCESSES_URI = f"{OGC_API_PROC_REL_BASE_URI}/processes"
+OGC_API_PROC_REL_PROCESS_DESC_URI = f"{OGC_API_PROC_REL_BASE_URI}/process-desc"
+OGC_API_PROC_REL_JOB_RESULTS_URI = f"{OGC_API_PROC_REL_BASE_URI}/results"
+OGC_API_PROC_REL_JOB_LIST_URI = f"{OGC_API_PROC_REL_BASE_URI}/job-list"
+OGC_API_PROC_REL_JOB_LOG_URI = f"{OGC_API_PROC_REL_BASE_URI}/log"
+
+OGC_API_PROC_PROFILE_BASE_URI = "https://www.opengis.net/def/profile/OGC/0"
+OGC_API_PROC_PROFILE_PROC_DESC_URI = f"{OGC_API_PROC_PROFILE_BASE_URI}/ogc-process-description"
+OGC_API_PROC_PROFILE_PROC_LIST_URI = f"{OGC_API_PROC_PROFILE_BASE_URI}/ogc-process-list"
+OGC_API_PROC_PROFILE_EXECUTE_URI = f"{OGC_API_PROC_PROFILE_BASE_URI}/ogc-execute-request"
+OGC_API_PROC_PROFILE_RESULTS_URI = f"{OGC_API_PROC_PROFILE_BASE_URI}/ogc-results"
+OGC_API_PROC_PROFILE_JOB_DESC_URI = f"{OGC_API_PROC_PROFILE_BASE_URI}/job-description"
+OGC_API_PROC_PROFILE_JOB_LIST_URI = f"{OGC_API_PROC_PROFILE_BASE_URI}/jobs-list"
+
+# Following are legacy definitions from OGC Testebeds and EO Apps Best Practices.
+# Therefore, leave them in 'http' for any operation relying on them explicitly (no auto https/versionless handling).
 OGC_API_PROC_PROFILE_DOCKER_APP_URI = "http://www.opengis.net/profiles/eoc/dockerizedApplication"
 OGC_API_PROC_PROFILE_WPS_APP_URI = "http://www.opengis.net/profiles/eoc/wpsApplication"
 
@@ -839,16 +862,39 @@ class AcceptLanguageHeader(ExtendedSchemaNode):
     default = AcceptLanguage.EN_CA  # FIXME: oneOf validator for supported languages (?)
 
 
-class AcceptProfileOneOf(OneOf):
+class AcceptWithProfileOneOf(OneOfCaseInsensitive):
+    """
+    Defines a ``Accept`` header that embeds a ``profile`` paremeter quoted accordingly (if URI) when provided as tuple.
+    """
+
+    def __init__(self, choices, *args, **kwargs):
+        # type: (Iterable[Union[str, Tuple[str, str]]], *Any, **Any) -> None
+        values = []
+        for choice in choices:
+            if isinstance(choice, tuple):
+                profile = choice[1]
+                profile = f"\"{profile}\"" if profile.split("://")[0] in ["http", "https"] else profile
+                values.append(f"{choice[0]}; profile={profile}")
+            else:
+                values.append(choice)
+        super().__init__(values, *args, **kwargs)
+
+
+class ProfileOneOf(OneOfCaseInsensitive):
+    """
+    Validates allowed values of an ``Accept-Profile`` header, ``profile`` query or other similar parameters.
+    """
+
     def __call__(self, node, value):
-        value = value.strip("<>")
+        value = value.strip("<>").strip("\"")
         return super().__call__(node, value)
 
 
 class AcceptProfileHeader(URI):
     name = "Accept-Profile"
     default = None
-    validator = AcceptProfileOneOf(
+    missing = drop
+    validator = ProfileOneOf(
         choices=[
             OGC_API_PROC_PROFILE_PROC_DESC_URI,
             OGC_API_PROC_PROFILE_PROC_LIST_URI,
@@ -858,7 +904,8 @@ class AcceptProfileHeader(URI):
             OGC_API_PROC_PROFILE_JOB_LIST_URI,
             OGC_WPS_1_SCHEMA_JOB_STATUS_URI,
             OPENEO_API_SCHEMA_JOB_STATUS_URI,
-        ]
+        ],
+        populate_variants=False,
     )
 
 
@@ -937,14 +984,14 @@ class RequestHeadersNoBody(ExtendedMappingSchema):
     accept_profile = AcceptProfileHeader(missing=drop)
 
 
-class RequestHeaders(RequestHeadersNoBody):
+class RequestHeadersBody(RequestHeadersNoBody):
     """
     Headers that can indicate how to adjust the behavior and/or result to be provided in the response.
     """
     content_type = RequestContentTypeHeader()
 
 
-class RequestHeadersAcceptAny(RequestHeaders):
+class RequestHeadersAcceptAny(RequestHeadersNoBody):
     """
     Headers that can indicate how to adjust the behavior and/or result to be provided in the response.
     """
@@ -1019,7 +1066,7 @@ class NoContent(ExtendedMappingSchema):
     default = {}
 
 
-class FileUploadHeaders(RequestHeaders):
+class FileUploadHeaders(RequestHeadersBody):
     # MUST be multipart for upload
     content_type = ContentTypeHeader(
         example=f"{ContentType.MULTIPART_FORM}; boundary=43003e2f205a180ace9cd34d98f911ff",
@@ -2415,7 +2462,7 @@ class JobStatusSearchEnum(ExtendedSchemaNode):
     title = "JobStatusSearch"
     default = Status.ACCEPTED
     example = Status.ACCEPTED
-    validator = StringOneOf(JOB_STATUS_SEARCH_API, delimiter=",", case_sensitive=False)
+    validator = DelimitedStringOneOf(JOB_STATUS_SEARCH_API, delimiter=",", case_sensitive=False)
 
 
 class JobTypeEnum(ExtendedSchemaNode):
@@ -2482,7 +2529,7 @@ class JobGroupsCommaSeparated(ExpandStringList, ExtendedSchemaNode):
     example = "process,service"
     missing = drop
     description = "Comma-separated list of grouping fields with which to list jobs."
-    validator = StringOneOf(["process", "provider", "service", "status"], delimiter=",", case_sensitive=True)
+    validator = DelimitedStringOneOf(["process", "provider", "service", "status"], delimiter=",", case_sensitive=True)
 
 
 class JobExecuteSubscribers(ExtendedMappingSchema):
@@ -2649,11 +2696,11 @@ class OutputPath(ExtendedMappingSchema):
 
 
 class FrontpageEndpoint(ExtendedMappingSchema):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
 
 
 class VersionsEndpoint(ExtendedMappingSchema):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
 
 
 class ConformanceQueries(ExtendedMappingSchema):
@@ -2667,7 +2714,7 @@ class ConformanceQueries(ExtendedMappingSchema):
 
 
 class ConformanceEndpoint(ExtendedMappingSchema):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = ConformanceQueries()
 
 
@@ -2677,7 +2724,7 @@ class OpenAPIAcceptHeader(AcceptHeader):
     validator = OneOf([ContentType.APP_OAS_JSON, ContentType.APP_JSON])
 
 
-class OpenAPIRequestHeaders(RequestHeaders):
+class OpenAPIRequestHeaders(RequestHeadersNoBody):
     accept = OpenAPIAcceptHeader()
 
 
@@ -3433,7 +3480,7 @@ class ErrorWPSResponse(ExtendedMappingSchema):
 
 
 class ProviderEndpoint(ProviderPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
 
 
 class ProcessDescriptionQuery(ExtendedMappingSchema):
@@ -3468,12 +3515,8 @@ class LocalProcessEndpointHeadersNoBody(AcceptFormatHeaders, RequestHeadersNoBod
     pass
 
 
-class LocalProcessEndpointHeaders(AcceptFormatHeaders, RequestHeaders):  # order important for descriptions to appear
-    pass
-
-
 class ProcessEndpoint(LocalProcessPath):
-    header = LocalProcessEndpointHeaders()
+    header = LocalProcessEndpointHeadersNoBody()
     querystring = LocalProcessDescriptionQuery()
 
 
@@ -3545,7 +3588,7 @@ class KVPInputQualifiedValue(ExtendedSchemaNode):
         "Examples:\n"
         "- Binary data: ``data[value]=SGVsbG8=&data[mediaType]=text/plain&data[encoding]=base64``\n"
         "- Structured data: ``geojson[value]={...}&geojson[mediaType]=application/geo+json"
-        "&geojson[profile]=http://www.opengis.net/spec/ogcapi-features-1/1.0``"
+        "&geojson[profile]=http://www.opengis.net/def/profile/ogc/0/rel-as-uri``"
     )
     example = "data_value"
     missing = drop
@@ -3622,9 +3665,9 @@ class KVPInputQualifiedProfile(ExtendedSchemaNode):
         "Replace ``{inputID}`` with the actual input ID. "
         "\n\n"
         "Example: ``data[value]={{...}}&data[mediaType]=application/geo+json"
-        "&data[profile]=http://www.opengis.net/spec/ogcapi-features-1/1.0``"
+        "&data[profile]=http://www.opengis.net/def/profile/ogc/0/rel-as-uri``"
     )
-    example = "http://www.opengis.net/spec/ogcapi-features-1/1.0"
+    example = "http://www.opengis.net/def/profile/ogc/0/rel-as-uri"
     missing = drop
 
 
@@ -3660,9 +3703,9 @@ class KVPInputBBoxCRS(ExtendedSchemaNode):
         "Used in conjunction with a bbox input parameter to specify the CRS. "
         "Replace ``{inputID}`` with the actual bbox input ID. "
         "\n\n"
-        "Example: ``bbox=5.8,47.2,15.1,55.1&bbox[crs]=http://www.opengis.net/def/crs/OGC/1.3/CRS84``"
+        f"Example: ``bbox=5.8,47.2,15.1,55.1&bbox[crs]={OGC_API_PROC_BBOX_CRS}``"
     )
-    example = "http://www.opengis.net/def/crs/OGC/1.3/CRS84"
+    example = OGC_API_PROC_BBOX_CRS
     missing = drop
 
 
@@ -3720,7 +3763,7 @@ class KVPOutputProfile(ExtendedSchemaNode):
         "Specifies a profile URI or well-known identifier for the format. "
         "Replace ``{inputID}`` or ``{outputID}`` with the actual identifier. "
         "\n\n"
-        "Example: ``data[profile]=http://www.opengis.net/spec/ogcapi-features-1/1.0``"
+        "Example: ``data[profile]=\"http://www.opengis.net/def/profile/ogc/0/rel-as-uri\"``"
     )
     missing = drop
 
@@ -3802,7 +3845,7 @@ class KVPResponseProfile(ExtendedSchemaNode, OAS3Parameter):
         "whereas the ``profile`` query applies to the "
         "[*Job Status*](https://pavics-weaver.readthedocs.io/en/latest/processes.html#alternate-job-status) response\n"
         "\n\n"
-        "Example: ``response[profile]=http://www.opengis.net/spec/ogcapi-features-1/1.0``"
+        f"Example: ``response[profile]=\"{OGC_API_PROC_PROFILE_JOB_DESC_URI}\"``"
     )
     missing = drop
 
@@ -3867,9 +3910,22 @@ class ProcessExecutionKVPInputOutputQuery(PermissiveMappingSchema, OAS3Parameter
         "KVPQualifiedInput": {
             "summary": "Qualified value with format specifications",
             "value": {
-                "data[value]": "{\"test\":123}",
-                "data[mediaType]": ContentType.APP_JSON,
-                "data[profile]": "http://www.opengis.net/spec/ogcapi-features-1/1.0",
+                # example from https://docs.ogc.org/DRAFTS/23-058r1.html#req_profile-references_rel-as-uri
+                "data[value]": json.dumps(
+                    {
+                        "type": "Feature",
+                        "id": 1,
+                        "geometry": {"type": "Point", "coordinates": [7.2789399, 50.7772485]},
+                        "properties": {
+                            "timeOfAccident": "2019-02-05T07:00:00Z",
+                            "roadSegment": "https://example.com/apis/roads/collections/roads/items/5209062A5209047O",
+                            "distanceFromStart": 851.0
+                        },
+                    },
+                    indent=0,
+                ),
+                "data[mediaType]": ContentType.APP_GEOJSON,
+                "data[profile]": "http://www.opengis.net/def/profile/ogc/0/rel-as-uri",
             }
         },
         "KVPByReference": {
@@ -3883,7 +3939,7 @@ class ProcessExecutionKVPInputOutputQuery(PermissiveMappingSchema, OAS3Parameter
             "summary": "Bounding box with CRS",
             "value": {
                 "bbox": "5.8,47.2,15.1,55.1",
-                "bbox[crs]": "http://www.opengis.net/def/crs/OGC/1.3/CRS84"
+                "bbox[crs]": OGC_API_PROC_BBOX_CRS,
             }
         },
         "KVPWithOutputs": {
@@ -3917,7 +3973,7 @@ class ProcessExecutionKVPInputOutputQuery(PermissiveMappingSchema, OAS3Parameter
                 "input1": "value1",
                 "response[f]": ContentType.APP_JSON,
                 "response[prefer]": ExecuteControlOption.ASYNC,
-                "response[profile]": "http://www.opengis.net/spec/ogcapi-processes-1/1.0",
+                "response[profile]": OGC_API_PROC_PROFILE_JOB_DESC_URI,
             }
         }
     }
@@ -3943,7 +3999,7 @@ class ProcessExecutionKVPEndpoint(LocalProcessPath):
 
 
 class ProcessPackageEndpoint(LocalProcessPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = LocalProcessQuery()
 
 
@@ -3952,35 +4008,45 @@ class ProviderProcessPackageEndpoint(ProviderProcessPath, ProcessPackageEndpoint
 
 
 class ProcessPayloadEndpoint(LocalProcessPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = LocalProcessQuery()
 
 
 class ProcessQuoteEstimatorGetEndpoint(LocalProcessPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = LocalProcessQuery()
 
 
 class ProcessQuoteEstimatorPutEndpoint(LocalProcessPath):
-    header = RequestHeaders()
+    header = RequestHeadersBody()
     querystring = LocalProcessQuery()
     body = QuoteEstimatorSchema()
 
 
 class ProcessQuoteEstimatorDeleteEndpoint(LocalProcessPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = LocalProcessQuery()
 
 
 class ProcessVisibilityGetEndpoint(LocalProcessPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = LocalProcessQuery()
 
 
 class ProcessVisibilityPutEndpoint(LocalProcessPath):
-    header = RequestHeaders()
+    header = RequestHeadersBody()
     querystring = LocalProcessQuery()
     body = VisibilitySchema()
+
+
+class GetJobAcceptProfileHeader(AcceptProfileHeader):
+    validator = ProfileOneOf(
+        choices=[
+            OGC_API_PROC_PROFILE_JOB_DESC_URI,
+            OGC_WPS_1_SCHEMA_JOB_STATUS_URI,
+            OPENEO_API_SCHEMA_JOB_STATUS_URI,
+        ]
+    )
 
 
 class JobStatusQueryProfileSchema(ExtendedSchemaNode):
@@ -3990,21 +4056,47 @@ class JobStatusQueryProfileSchema(ExtendedSchemaNode):
     title = "JobStatusQuerySchema"
     example = JobStatusProfileSchema.OGC
     default = JobStatusProfileSchema.OGC
-    validator = OneOfCaseInsensitive(JobStatusProfileSchema.values())
+    validator = ProfileOneOf(JobStatusProfileSchema.values() + GetJobAcceptProfileHeader.validator.choices)
 
 
 class GetJobQuery(FormatQuery):
-    schema = JobStatusQueryProfileSchema(missing=drop)
+    schema = JobStatusQueryProfileSchema(missing=drop, default=None, deprecated=True)
     profile = JobStatusQueryProfileSchema(missing=drop)
 
 
+class GetJobAcceptHeaderJSON(AcceptHeader):
+    validator = AcceptWithProfileOneOf(
+        choices=(
+            # JSON by itself or embedding a 'profile' as shortcut or full URI
+            [ContentType.APP_JSON] +
+            [
+                # note:
+                #   typically non-JSON like interfacs like WPS (XML) are still allowed
+                #   in the context of Job Status response, it means to get the OGC API JSON with WPS status values
+                #   (i.e.: the 'WPS' portion refers to the 'weaver.status.map_status' resolution)
+                (ContentType.APP_JSON, profile)
+                for profile in JobStatusQueryProfileSchema.validator.choices
+            ]
+        ),
+    )
+
+
+class GetJobHeaders(RequestHeadersNoBody):
+    # This endpoint offers the representations across media-types/profiles.
+    # Therefore, although only certain combinations are allowed, they must all be present to render them in OpenAPI.
+    # Also, because we offer the 'Accept-Profile' header and 'profile' query, we purposely omit (though still allowed)
+    # the variant of 'Accept' with embedded 'profile' to keep definitions more concise and easier to employ by users.
+    accept = AcceptHeader()
+    accept_profile = GetJobAcceptProfileHeader()
+
+
 class GetProviderJobEndpoint(ProviderProcessPath, JobPath):
-    header = RequestHeaders()
+    header = GetJobHeaders()
     querystring = GetJobQuery()
 
 
 class GetJobEndpoint(JobPath):
-    header = RequestHeaders()
+    header = GetJobHeaders()
     querystring = GetJobQuery()
 
 
@@ -4059,31 +4151,36 @@ class JobInputsOutputsQuery(ExtendedMappingSchema):
 
 
 class ProcessJobDefinitionEndpoint(LocalProcessPath, JobPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
+    querystring = JobInputsOutputsQuery()
 
 
 class ProviderJobDefinitionEndpoint(ProviderProcessPath, JobPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
+    querystring = JobInputsOutputsQuery()
 
 
 class JobDefinitionEndpoint(JobPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = JobInputsOutputsQuery()
 
 
 class ProcessJobInputsEndpoint(LocalProcessPath, JobPath):
     deprecated = True
     header = RequestHeadersNoBody()
+    querystring = JobInputsOutputsQuery()
 
 
 class ProviderJobInputsEndpoint(ProviderProcessPath, JobPath):
     deprecated = True
     header = RequestHeadersNoBody()
+    querystring = JobInputsOutputsQuery()
 
 
 class JobInputsEndpoint(JobPath):
     deprecated = True
     header = RequestHeadersNoBody()
+    querystring = JobInputsOutputsQuery()
 
 
 class JobResultsQuery(FormatQuery):
@@ -4117,7 +4214,7 @@ class JobOutputsEndpoint(JobPath):
 
 
 class JobOutputEndpoint(JobPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = LocalProcessJobResultsQuery()
 
 
@@ -4127,7 +4224,7 @@ class ProcessOutputsEndpoint(LocalProcessPath, JobPath):
 
 
 class ProcessOutputEndpoint(LocalProcessPath, JobPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = LocalProcessJobResultsQuery()
 
 
@@ -4137,7 +4234,7 @@ class ProviderOutputsEndpoint(ProviderProcessPath, JobPath):
 
 
 class ProviderOutputEndpoint(ProviderProcessPath, JobPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = JobResultsQuery()
 
 
@@ -4193,7 +4290,7 @@ class ProviderResultsIndexEndpoint(ProviderProcessPath, JobResultsIndexPath):
 
 
 class JobResultsTriggerExecutionEndpoint(JobResultsEndpoint):
-    header = RequestHeaders()
+    header = RequestHeadersBody()
     body = NoContent()
 
 
@@ -7492,7 +7589,7 @@ class Deploy(OneOfKeywordSchema):
     ]
 
 
-class DeployContentType(ContentTypeHeader):
+class DeployContentTypeAny(ContentTypeHeader):
     example = ContentType.APP_JSON
     default = ContentType.APP_JSON
     validator = OneOf([
@@ -7507,19 +7604,52 @@ class DeployContentType(ContentTypeHeader):
     ])
 
 
-class DeployHeaders(RequestHeaders):
+class DeployContentTypeOGC(ContentTypeHeader):
+    example = ContentType.APP_JSON
+    default = ContentType.APP_JSON
+    validator = OneOf([
+        ContentType.APP_JSON,
+        ContentType.APP_OGC_PKG_JSON,
+        ContentType.APP_OGC_PKG_YAML,
+        ContentType.APP_YAML,
+    ])
+
+
+class DeployContentTypeCWL(ContentTypeHeader):
+    example = ContentType.APP_JSON
+    default = ContentType.APP_JSON
+    validator = OneOf([
+        ContentType.APP_CWL,
+        ContentType.APP_CWL_JSON,
+        ContentType.APP_CWL_YAML,
+        ContentType.APP_CWL_X,
+        ContentType.APP_YAML,
+        ContentType.APP_JSON,
+    ])
+
+
+class DeployHeadersAny(RequestHeadersBody):
     x_auth_docker = XAuthDockerHeader()
-    content_type = DeployContentType()
+    content_type = DeployContentTypeAny()
+
+
+class DeployHeadersOGC(RequestHeadersBody):
+    x_auth_docker = XAuthDockerHeader()
+    content_type = DeployContentTypeOGC()
+
+
+class DeployHeadersCWL(RequestHeadersBody):
+    x_auth_docker = XAuthDockerHeader()
+    content_type = DeployContentTypeCWL()
 
 
 class PostProcessesEndpoint(ExtendedMappingSchema):
-    header = DeployHeaders(description="Headers employed for process deployment.")
+    header = DeployHeadersAny(description="Headers employed for process deployment.")
     querystring = FormatQuery()
-    body = Deploy(title="Deploy", examples={
-        "DeployCWL": {
-            "summary": "Deploy a process from a CWL+JSON definition.",
-            "value": EXAMPLES["deploy_process_cwl.json"],
-        },
+
+
+class DeployBodyOGC(Deploy):
+    examples = {
         "DeployOGC": {
             "summary": "Deploy a process from an OGC Application Package definition.",
             "value": EXAMPLES["deploy_process_ogcapppkg.json"],
@@ -7528,16 +7658,36 @@ class PostProcessesEndpoint(ExtendedMappingSchema):
             "summary": "Deploy a process from a remote WPS-1 reference URL.",
             "value": EXAMPLES["deploy_process_wps1.json"],
         }
-    })
+    }
 
 
-class PostProcessesEndpointCWLYAML(PostProcessesEndpoint):
-    body = PermissiveMappingSchema(title="DeployCWLYAML", examples={
+class DeployBodyCWL(Deploy):
+    examples = {
+        "DeployCWLJSON": {
+            "summary": "Deploy a process from a CWL+JSON definition.",
+            "value": EXAMPLES["deploy_process_cwl.json"],
+        },
         "DeployCWLYAML": {
             "summary": "Deploy a process from a CWL+YAML definition.",
             "value": EXAMPLES["deploy_process_yaml.cwl"],
         },
-    })
+    }
+
+
+class PostProcessesEndpointOGC(PostProcessesEndpoint):
+    header = DeployHeadersOGC(
+        description="Headers employed for process deployment to represent OGC Application Package content.",
+    )
+    querystring = FormatQuery()
+    body = DeployBodyOGC()
+
+
+class PostProcessesEndpointCWL(PostProcessesEndpoint):
+    header = DeployHeadersCWL(
+        description="Headers employed for process deployment to represent CWL content.",
+    )
+    querystring = FormatQuery()
+    body = DeployBodyCWL()
 
 
 class UpdateInputOutputBase(DescriptionType, InputOutputDescriptionMeta):
@@ -7627,18 +7777,18 @@ class PutProcessBodySchema(Deploy):
 
 
 class PatchProcessEndpoint(LocalProcessPath):
-    headers = RequestHeaders()
+    headers = RequestHeadersBody()
     querystring = LocalProcessQuery()
     body = PatchProcessBodySchema()
 
 
 class PutProcessEndpoint(LocalProcessPath):
-    headers = RequestHeaders()
+    headers = RequestHeadersBody()
     querystring = LocalProcessQuery()
     body = PutProcessBodySchema()
 
 
-class ExecuteHeadersBase(RequestHeaders):
+class ExecuteHeadersBase(RequestHeadersBody):
     description = "Request headers supported for job execution."
     prefer = PreferHeader(missing=drop)
     x_wps_output_context = WpsOutputContextHeader()
@@ -7733,7 +7883,7 @@ class PatchProcessJobEndpoint(JobPath, ProcessEndpoint):
 
 
 class PatchProviderJobEndpoint(PatchProcessJobEndpoint):
-    header = RequestHeaders()
+    header = RequestHeadersBody()
 
 
 class PagingQueries(ExtendedMappingSchema):
@@ -7789,17 +7939,17 @@ class GetProviderJobsQueries(GetJobsQueries):  # ':version' not allowed for proc
 
 
 class GetJobsEndpoint(ExtendedMappingSchema):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = GetProcessJobsQuery()  # allowed version in this case since can be either local or remote processes
 
 
 class GetProcessJobsEndpoint(LocalProcessPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = GetProcessJobsQuery()
 
 
 class GetProviderJobsEndpoint(ProviderProcessPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = GetProviderJobsQueries()
 
 
@@ -7812,7 +7962,7 @@ class DeleteJobsBodySchema(ExtendedMappingSchema):
 
 
 class DeleteJobsEndpoint(ExtendedMappingSchema):
-    header = RequestHeaders()
+    header = RequestHeadersBody()
     body = DeleteJobsBodySchema()
 
 
@@ -7828,18 +7978,18 @@ class GetProcessJobQuery(LocalProcessQuery, GetJobQuery):
     pass
 
 
-class GetProcessJobEndpoint(LocalProcessPath):
-    header = RequestHeaders()
+class GetProcessJobEndpoint(LocalProcessPath, JobPath):
+    header = RequestHeadersNoBody()
     querystring = GetProcessJobQuery()
 
 
 class DeleteJobEndpoint(JobPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = LocalProcessQuery()
 
 
 class DeleteProcessJobEndpoint(LocalProcessPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
 
 
 class DeleteProviderJobEndpoint(DeleteProcessJobEndpoint, ProviderProcessPath):
@@ -7847,20 +7997,20 @@ class DeleteProviderJobEndpoint(DeleteProcessJobEndpoint, ProviderProcessPath):
 
 
 class BillsEndpoint(ExtendedMappingSchema):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
 
 
 class BillEndpoint(BillPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
 
 
 class ProcessQuotesEndpoint(LocalProcessPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = LocalProcessQuery()
 
 
 class ProcessQuoteEndpoint(LocalProcessPath, QuotePath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = LocalProcessQuery()
 
 
@@ -7870,22 +8020,22 @@ class GetQuotesQueries(PagingQueries):
 
 
 class QuotesEndpoint(ExtendedMappingSchema):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = GetQuotesQueries()
 
 
 class QuoteEndpoint(QuotePath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
 
 
 class PostProcessQuote(LocalProcessPath, QuotePath):
-    header = RequestHeaders()
+    header = RequestHeadersBody()
     querystring = LocalProcessQuery()
     body = NoContent()
 
 
 class PostQuote(QuotePath):
-    header = RequestHeaders()
+    header = RequestHeadersBody()
     body = NoContent()
 
 
@@ -7894,7 +8044,7 @@ class QuoteProcessParametersSchema(ExecuteInputOutputs):
 
 
 class PostProcessQuoteRequestEndpoint(LocalProcessPath, QuotePath):
-    header = RequestHeaders()
+    header = RequestHeadersBody()
     querystring = LocalProcessQuery()
     body = QuoteProcessParametersSchema()
 
@@ -7926,11 +8076,11 @@ class ProvidersQuerySchema(ExtendedMappingSchema):
 
 class GetProviders(ExtendedMappingSchema):
     querystring = ProvidersQuerySchema()
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
 
 
 class PostProvider(ExtendedMappingSchema):
-    header = RequestHeaders()
+    header = RequestHeadersBody()
     body = CreateProviderRequestBody()
 
 
@@ -7968,12 +8118,12 @@ class ProviderProcessesQuery(ProcessPagingQuery, ProcessDetailQuery, ProcessLink
 
 
 class ProviderProcessesEndpoint(ProviderPath):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
     querystring = ProviderProcessesQuery()
 
 
 class GetProviderProcess(ExtendedMappingSchema):
-    header = RequestHeaders()
+    header = RequestHeadersNoBody()
 
 
 class PostProviderProcessJobRequest(ExtendedMappingSchema):
@@ -8523,7 +8673,7 @@ class JobProvAcceptHeader(AcceptHeader):
     validator = OneOf(ProvenanceFormat.media_types())
 
 
-class JobProvRequestHeaders(RequestHeaders):
+class JobProvRequestHeaders(RequestHeadersNoBody):
     accept = JobProvAcceptHeader()
 
 
@@ -8729,10 +8879,6 @@ class VaultAccessToken(UUID):
     example = "30d889cfb7ae3a63229a8de5f91abc1ef5966bb664972f234a4db9d28f8148e0e"  # nosec
 
 
-class VaultEndpoint(ExtendedMappingSchema):
-    header = RequestHeaders()
-
-
 class VaultUploadBody(ExtendedSchemaNode):
     schema_type = String
     description = "Multipart file contents for upload to the vault."
@@ -8800,7 +8946,7 @@ class VaultFileRequestHeaders(ExtendedMappingSchema):
     access_token = XAuthVaultFileHeader()
 
 
-class VaultFileEndpoint(VaultEndpoint):
+class VaultFileEndpoint(ExtendedMappingSchema):
     header = VaultFileRequestHeaders()
     file_id = VaultFileID()
 
@@ -9233,12 +9379,24 @@ get_result_redirect_responses = {
     "308": RedirectResponse(description="Redirects '/result' (without 's') to corresponding '/results' path."),
 }
 get_job_results_responses = {
-    "200": OkGetJobResultsResponse(description="success", examples={
-        "JobResults": {
-            "summary": "Obtained job results.",
-            "value": EXAMPLES["job_results.json"],
+    "200": OkGetJobResultsResponse(
+        description=(
+            "Retrieval of results from a successfully completed job."
+            "\n\n"
+            "For bacward compatibility, nested `format` content details are provided. "
+            "However, implementations should rely on directly provided `mediaType`, `encoding`, etc."
+        ),
+        examples={
+            "JobResultsByReference": {
+                "summary": "Obtained job results by reference.",
+                "value": EXAMPLES["job_results_by_reference.json"],
+            },
+            "JobResultsByValue": {
+                "summary": "Obtained job results by value.",
+                "value": EXAMPLES["job_results_by_value.json"],
+            }
         }
-    }),
+    ),
     "204": NoContentJobResultsResponse(description="success"),
     "400": InvalidJobResponseSchema(),
     "404": NotFoundJobResponseSchema(),